requirements.txt
ccxt
pandas
pytz
requests
supabase-py
python-telegram-bot==21.0.1